+++ src/components/base.js
@@ -1,8 +1,8 @@
 //@flow
 import {Component} from 'preact';
-import Player from 'playkit-js';
 import getLogger from '../utils/logger';
 import {types} from '../middlewares/event-dispatcher';
+import {EventManager} from '../event/event-manager';
 
 /**
  * Base component to be extended by other player UI components
@@ -12,10 +12,11 @@ import {types} from '../middlewares/event-dispatcher';
  */
 class BaseComponent extends Component {
   state: Object;
-  player: Player;
+  player: Object;
   name: string;
   config: Object;
   logger: any;
+  eventManager: EventManager;
 
   /**
    * Components default props.
@@ -34,6 +35,7 @@ class BaseComponent extends Component {
     this.name = obj.name;
     this.player = obj.player;
     this.config = obj.config;
+    this.eventManager = new EventManager();
     this.logger = getLogger(`UI ${this.name}`);
     this.logger.debug(`Initialized`);
   }
@@ -66,6 +68,16 @@ class BaseComponent extends Component {
       name: this.name,
       payload: payload
     });
+  }
+
+  /**
+   * Before component is mounted remove all event manager listeners.
+   * @returns {void}
+   *
+   * @memberof BaseComponent
+   */
+  componentWillUnmount(): void {
+    this.eventManager.removeAll();
   }
 }