+++ src/components/engine-connector/engine-connector.js
@@ -1,13 +1,16 @@
 //@flow
 import {h} from 'preact';
+import style from '../../styles/style.scss';
 import {connect} from 'preact-redux';
 import {bindActions} from '../../utils/bind-actions';
 import {default as reduce, actions} from '../../reducers/engine';
+import {actions as loadingActions} from '../../reducers/loading';
+import {actions as shellActions} from '../../reducers/shell';
 import BaseComponent from '../base';
 import {EventManager} from '../../event/event-manager';
 import {UIEventManager} from '../../event/event-manager';
 
-@connect(reduce, bindActions(actions))
+@connect(reduce, bindActions({...actions, ...loadingActions, ...shellActions}))
   /**
    * EngineConnector component
    *
@@ -37,43 +40,50 @@ class EngineConnector extends BaseComponent {
   componentDidMount() {
     const TrackType = this.player.Track;
 
-    this._eventManager.listen(this.player, this.player.Event.PLAYER_RESET, () => {
+    this.eventManager.listen(this.player, this.player.Event.PLAYER_RESET, () => {
       this.props.updateIsIdle(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.SOURCE_SELECTED, () => {
+    this.eventManager.listen(this.player, this.player.Event.SOURCE_SELECTED, () => {
       this.props.updateIsVr(this.player.isVr());
+      if (this.player.config.playback.autoplay) {
+        this.props.updateLoadingSpinnerState(true);
+      } else {
+        this.props.updateLoadingSpinnerState(false);
+        this.props.updatePrePlayback(true);
+        this.props.addPlayerClass(style.prePlayback);
+      }
     });
 
-    this._eventManager.listen(this.player, this.player.Event.CHANGE_SOURCE_STARTED, () => {
+    this.eventManager.listen(this.player, this.player.Event.CHANGE_SOURCE_STARTED, () => {
       this.props.updateFallbackToMutedAutoPlay(false);
       this.props.updateAdBreak(false);
       this.props.updateAdIsPlaying(false);
       this.props.updateIsPlaying(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.CHANGE_SOURCE_ENDED, () => {
+    this.eventManager.listen(this.player, this.player.Event.CHANGE_SOURCE_ENDED, () => {
       this.props.updatePlayerPoster(this.player.poster);
       this.props.updateIsIdle(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.PLAYER_STATE_CHANGED, (e) => {
+    this.eventManager.listen(this.player, this.player.Event.PLAYER_STATE_CHANGED, (e) => {
       this.props.updatePlayerState(e.payload.oldState.type, e.payload.newState.type);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.TIME_UPDATE, () => {
+    this.eventManager.listen(this.player, this.player.Event.TIME_UPDATE, () => {
       this.props.updateCurrentTime(this.player.currentTime);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.DURATION_CHANGE, () => {
+    this.eventManager.listen(this.player, this.player.Event.DURATION_CHANGE, () => {
       this.props.updateDuration(this.player.duration);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.LOADED_DATA, () => {
+    this.eventManager.listen(this.player, this.player.Event.LOADED_DATA, () => {
       this.props.updateDuration(this.player.duration);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.LOADED_METADATA, () => {
+    this.eventManager.listen(this.player, this.player.Event.LOADED_METADATA, () => {
       this.props.updateMuted(this.player.muted);
       this.props.updateMetadataLoadingStatus(true);
       this.props.updateIsLive(this.player.isLive());
@@ -81,15 +91,15 @@ class EngineConnector extends BaseComponent {
       this.props.updatePlayerPoster(this.player.poster);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.VOLUME_CHANGE, () => {
+    this.eventManager.listen(this.player, this.player.Event.VOLUME_CHANGE, () => {
       this.props.updateVolume(this.player.volume);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.MUTE_CHANGE, () => {
+    this.eventManager.listen(this.player, this.player.Event.MUTE_CHANGE, () => {
       this.props.updateMuted(this.player.muted);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.PLAYING, () => {
+    this.eventManager.listen(this.player, this.player.Event.PLAYING, () => {
       this.props.updateIsPlaying(true);
 
       if (this.props.engine.isEnded) {
@@ -97,15 +107,15 @@ class EngineConnector extends BaseComponent {
       }
     });
 
-    this._eventManager.listen(this.player, this.player.Event.PAUSE, () => {
+    this.eventManager.listen(this.player, this.player.Event.PAUSE, () => {
       this.props.updateIsPlaying(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.ENDED, () => {
+    this.eventManager.listen(this.player, this.player.Event.ENDED, () => {
       this.props.updateIsEnded(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.TRACKS_CHANGED, () => {
+    this.eventManager.listen(this.player, this.player.Event.TRACKS_CHANGED, () => {
       let audioTracks = this.player.getTracks(TrackType.AUDIO);
       let videoTracks = this.player.getTracks(TrackType.VIDEO);
       let textTracks = this.player.getTracks(TrackType.TEXT);
@@ -115,95 +125,95 @@ class EngineConnector extends BaseComponent {
       this.props.updateTextTracks(textTracks);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.TEXT_TRACK_CHANGED, () => {
+    this.eventManager.listen(this.player, this.player.Event.TEXT_TRACK_CHANGED, () => {
       let tracks = this.player.getTracks(TrackType.TEXT);
       this.props.updateTextTracks(tracks);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AUDIO_TRACK_CHANGED, () => {
+    this.eventManager.listen(this.player, this.player.Event.AUDIO_TRACK_CHANGED, () => {
       let tracks = this.player.getTracks(TrackType.AUDIO);
       this.props.updateAudioTracks(tracks);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.VIDEO_TRACK_CHANGED, () => {
+    this.eventManager.listen(this.player, this.player.Event.VIDEO_TRACK_CHANGED, () => {
       let tracks = this.player.getTracks(TrackType.VIDEO);
       this.props.updateVideoTracks(tracks);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_BREAK_START, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_BREAK_START, () => {
       this.props.updateHasError(false);
       this.props.updateAdBreak(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_BREAK_END, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_BREAK_END, () => {
       this.props.updateAdBreak(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.ALL_ADS_COMPLETED, () => {
+    this.eventManager.listen(this.player, this.player.Event.ALL_ADS_COMPLETED, () => {
       this.props.updateAdBreak(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_PROGRESS, e => {
+    this.eventManager.listen(this.player, this.player.Event.AD_PROGRESS, e => {
       let currentTime = e.payload.adProgress.currentTime;
       let duration = e.payload.adProgress.duration;
 
       this.props.updateAdBreakProgress(currentTime, duration);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_COMPLETED, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_COMPLETED, () => {
       this.props.updateAdBreakCompleted();
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_STARTED, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_STARTED, () => {
       this.props.updateAdIsPlaying(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_RESUMED, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_RESUMED, () => {
       this.props.updateAdIsPlaying(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_PAUSED, () => {
+    this.eventManager.listen(this.player, this.player.Event.AD_PAUSED, () => {
       this.props.updateAdIsPlaying(false);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_ERROR, e => {
+    this.eventManager.listen(this.player, this.player.Event.AD_ERROR, e => {
       if (e.payload.fatal) {
         this.props.updateAdBreak(false);
       }
     });
 
-    this._eventManager.listen(this.player, this.player.Event.FALLBACK_TO_MUTED_AUTOPLAY, () => {
+    this.eventManager.listen(this.player, this.player.Event.FALLBACK_TO_MUTED_AUTOPLAY, () => {
       this.props.updateFallbackToMutedAutoPlay(true);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.AD_LOADED, e => {
+    this.eventManager.listen(this.player, this.player.Event.AD_LOADED, e => {
       this.props.updateAdIsLinear(e.payload.ad.isLinear());
       this.props.updateAdClickUrl(e.payload.ad.g.clickThroughUrl);
       this.props.updateAdSkipTimeOffset(e.payload.ad.getSkipTimeOffset());
       this.props.updateAdSkippableState(e.payload.ad.getAdSkippableState());
     });
 
-    this._eventManager.listen(this.player, this.player.Event.VR_STEREO_MODE_CHANGED, e => {
+    this.eventManager.listen(this.player, this.player.Event.VR_STEREO_MODE_CHANGED, e => {
       this.props.updateVrStereoMode(e.payload.mode);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.ERROR, e => {
+    this.eventManager.listen(this.player, this.player.Event.ERROR, e => {
       if (e.payload && e.payload.severity === 2) {
         this.props.updateHasError(true);
       }
     });
 
-    this._eventManager.listen(this.player, this.player.Event.Cast.CAST_SESSION_STARTED, e => {
+    this.eventManager.listen(this.player, this.player.Event.Cast.CAST_SESSION_STARTED, e => {
       this.props.updateIsCasting(true);
       this.props.updateCastSession(e.payload.session);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.Cast.CAST_SESSION_ENDED, () => {
+    this.eventManager.listen(this.player, this.player.Event.Cast.CAST_SESSION_ENDED, () => {
       this.props.updateIsCasting(false);
       this.props.updateCastSession(null);
     });
 
-    this._eventManager.listen(this.player, this.player.Event.Cast.CAST_AVAILABLE, e => {
+    this.eventManager.listen(this.player, this.player.Event.Cast.CAST_AVAILABLE, e => {
       const {available, type} = e.payload;
       this.props.updateIsCastAvailable(available);
       if (available) {
@@ -229,7 +239,7 @@ class EngineConnector extends BaseComponent {
    * @memberof EngineConnector
    */
   render(): React$Element<any> {
-    return <span/>
+    return <span/>;
   }
 }