+++ src/components/pre-playback-play-overlay/pre-playback-play-overlay.js
@@ -8,8 +8,6 @@ import BaseComponent from '../base';
 import {default as Icon, IconType} from '../icon';
 import {KeyMap} from '../../utils/key-map';
 import {actions as loadingActions} from '../../reducers/loading';
-import {EventManager} from '../../event/event-manager';
-import {UIEventManager} from '../../event/event-manager';
 
 /**
  * mapping state to props
@@ -25,7 +23,7 @@ const mapStateToProps = state => ({
   loading: state.loading.show
 });
 
-@connect(mapStateToProps, bindActions(Object.assign(actions, loadingActions)))
+@connect(mapStateToProps, bindActions(Object.assign({}, actions, loadingActions)))
   /**
    * PrePlaybackPlayOverlay component
    *
@@ -35,7 +33,6 @@ const mapStateToProps = state => ({
    */
 class PrePlaybackPlayOverlay extends BaseComponent {
   autoplay: boolean;
-  _eventManager: EventManager;
 
   /**
    * Creates an instance of PrePlaybackPlayOverlay.
@@ -44,8 +41,6 @@ class PrePlaybackPlayOverlay extends BaseComponent {
    */
   constructor(obj: Object) {
     super({name: 'PrePlaybackPlayOverlay', player: obj.player});
-    this._eventManager = UIEventManager.getInstance();
-    this._eventManager.listen(this.player, this.player.Event.CHANGE_SOURCE_ENDED, () => this._onChangeSourceEnded());
   }
 
   /**
@@ -60,7 +55,7 @@ class PrePlaybackPlayOverlay extends BaseComponent {
     try {
       this.autoplay = this.player.config.playback.autoplay;
       if (this.autoplay === true) {
-        this._eventManager.listen(this.player, this.player.Event.AUTOPLAY_FAILED, () => {
+        this.eventManager.listen(this.player, this.player.Event.AUTOPLAY_FAILED, () => {
           this.autoplay = false;
           this.forceUpdate();
         });
@@ -77,6 +72,7 @@ class PrePlaybackPlayOverlay extends BaseComponent {
    * @memberof PrePlaybackPlayOverlay
    */
   componentWillUnmount() {
+    super.componentWillUnmount();
     this._hidePrePlayback();
     this.props.removePlayerClass(style.prePlayback);
   }
@@ -88,7 +84,7 @@ class PrePlaybackPlayOverlay extends BaseComponent {
    * @memberof PrePlaybackPlayOverlay
    */
   componentDidMount() {
-    this._eventManager.listen(this.player, this.player.Event.PLAY, () => this._hidePrePlayback());
+    this.eventManager.listen(this.player, this.player.Event.PLAY, () => this._hidePrePlayback());
     if (this.player.paused === false) {
       this._hidePrePlayback();
     }
@@ -140,7 +136,7 @@ class PrePlaybackPlayOverlay extends BaseComponent {
 
     if (!props.prePlayback && props.poster) {
       rootStyle = {backgroundImage: `url(${props.poster})`};
-      rootClass.push(style.hasPoster)
+      rootClass.push(style.hasPoster);
     }
 
     return (
@@ -159,32 +155,7 @@ class PrePlaybackPlayOverlay extends BaseComponent {
           {props.isEnded ? <Icon type={IconType.StartOver}/> : <Icon type={IconType.Play}/>}
         </a>}
       </div>
-    )
-  }
-
-  /**
-   * Change source ended event handler.
-   * @private
-   * @returns {void}
-   */
-  _onChangeSourceEnded(): void {
-    try {
-      if (!this.player.config.playback.autoplay) {
-        this._displayPrePlayback();
-      }
-    } catch (e) {
-      this.logger.error(e.message);
-    }
-  }
-
-  /**
-   * Displays the pre playback overlay.
-   * @private
-   * @returns {void}
-   */
-  _displayPrePlayback(): void {
-    this.props.updatePrePlayback(true);
-    this.props.addPlayerClass(style.prePlayback);
+    );
   }
 
   /**