+++ src/ui-manager.js
@@ -4,15 +4,13 @@ import {Provider} from 'preact-redux';
 import {IntlProvider} from 'preact-i18n';
 import {createStore} from 'redux';
 import {copyDeep} from './utils/copy-deep';
-import {LogLevel, getLogLevel, setLogLevel} from './utils/logger'
-import {EventType} from './event/event-type'
+import {LogLevel, getLogLevel, setLogLevel} from './utils/logger';
+import {EventType} from './event/event-type';
 
 import reducer from './store';
 import definition from './fr.json';
 
 import {actions} from './reducers/config';
-
-import {Player} from 'playkit-js';
 // core components for the UI
 import {EngineConnector} from './components/engine-connector';
 import {Shell} from './components/shell';
@@ -21,22 +19,21 @@ import {PlayerGUI} from './player-gui';
 // ui presets
 import * as presets from './ui-presets';
 
-import {middleware} from './middlewares'
+import {middleware} from './middlewares';
 
 import './styles/style.scss';
-import {EventManager} from './event/event-manager';
-import {UIEventManager} from './event/event-manager';
 
 /**
  * API used for building UIs based on state conditions
  *
  * @class UIManager
  */
-export default class UIManager {
-  player: Player;
+class UIManager {
+  player: Object;
   targetId: string;
   store: any;
-  _eventManager: EventManager;
+  container: ?HTMLElement;
+  root: React$Component<any, any, any>;
 
   /**
    * Creates an instance of UIManager.
@@ -50,7 +47,6 @@ export default class UIManager {
     }
     this.player = player;
     this.targetId = config.targetId;
-    this._eventManager = UIEventManager.getInstance();
     this._createStore(config);
     this.setConfig(config);
   }
@@ -137,22 +133,24 @@ export default class UIManager {
    */
   _buildUI(uis?: Array<UIPreset>): void {
     if (!this.player) return;
-    const container = document.getElementById(this.targetId);
-    // i18n, redux and initial player-to-store connector setup
-    const template = (
-      <Provider store={this.store}>
-        <IntlProvider definition={definition}>
-          <Shell player={this.player}>
-            <EngineConnector player={this.player}/>
-            <VideoPlayer player={this.player}/>
-            <PlayerGUI uis={uis} player={this.player} playerContainer={container}/>
-          </Shell>
-        </IntlProvider>
-      </Provider>
-    );
+    this.container = document.getElementById(this.targetId);
+    if (this.container) {
+      // i18n, redux and initial player-to-store connector setup
+      const template = (
+        <Provider store={this.store}>
+          <IntlProvider definition={definition}>
+            <Shell player={this.player}>
+              <EngineConnector player={this.player}/>
+              <VideoPlayer player={this.player}/>
+              <PlayerGUI uis={uis} player={this.player} playerContainer={this.container}/>
+            </Shell>
+          </IntlProvider>
+        </Provider>
+      );
 
-    // render the player
-    render(template, container);
+      // render the player
+      this.root = render(template, this.container);
+    }
   }
 
   /**
@@ -160,11 +158,10 @@ export default class UIManager {
    * @returns {void}
    */
   destroy(): void {
-    const container = document.getElementById(this.targetId);
-    if (container && container.childNodes) {
-      container.removeChild(container.childNodes[0]);
-      container.prepend(this.player.getView());
-      this._eventManager.removeAll();
+    // $FlowFixMe
+    render('', this.container, this.root);
+    if (this.container) {
+      this.container.prepend(this.player.getView());
     }
   }
 
@@ -192,7 +189,7 @@ export default class UIManager {
    * @returns {Object} - The log levels of the player.
    * @public
    */
-  get LogLevel(): { [level: string]: Object } {
+  get LogLevel(): {[level: string]: Object} {
     return LogLevel;
   }
 
@@ -201,7 +198,9 @@ export default class UIManager {
    * @returns {Object} - The ui manager event enums.
    * @public
    */
-  get Event(): { [event: string]: string } {
+  get Event(): {[event: string]: string} {
     return EventType;
   }
 }
+
+export {UIManager};